rename tests/common.h to tests/common.hpp
