+++ tests/treemap.cpp
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "common.hpp"
 
 #include <catch2/catch.hpp>
 #include <llama/llama.hpp>
